+++ terra_opp/settings.py
@@ -21,17 +21,9 @@ TROPP_VIEWPOINT_PROPERTIES_SET = {
     'filter': {},
 }
 
-TROPP_FEATURES_PROPERTIES_FROM_VIEWPOINT = ("commune",)
+TROPP_FEATURES_PROPERTIES_FROM_VIEWPOINT = ()
 
 TROPP_SEARCHABLE_PROPERTIES = {
-    'cities': {
-        'json_key': 'commune',
-        'type': 'single',
-    },
-    'themes': {
-        'json_key': 'themes',
-        'type': 'many',
-    },
     'road': {
         'json_key': 'voie',
         'type': 'text',